server/dev-requirements.txt
altgraph==0.17.5
bottle==0.13.4
click==8.3.1
colorama==0.4.6
h11==0.16.0
macholib==1.16.4
packaging==26.0
proxy_tools==0.1.0
pyinstaller==6.19.0
pyinstaller-hooks-contrib==2026.3
pyobjc-core==12.1
pyobjc-framework-Cocoa==12.1
pyobjc-framework-Quartz==12.1
pyobjc-framework-Security==12.1
pyobjc-framework-UniformTypeIdentifiers==12.1
pyobjc-framework-WebKit==12.1
pywebview==6.1
setuptools==82.0.1
typing_extensions==4.15.0
uvicorn==0.41.0
websockets==15.0.1
pyserial==3.5
pynmea2==1.19.0
